refactor: 입출력 메시지 분리
