+++ src/constants/io-message.js
@@ -1,0 +1,12 @@
+const INPUT_MESSAGE = Object.freeze({
+  RACING_CARS:
+    "경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)\n",
+  ATTEMPT_COUNT: "시도할 횟수는 몇 회인가요?\n",
+});
+
+const OUTPUT_MESSAGE = Object.freeze({
+  ROUND_RESULT: "\n실행 결과",
+  FINAL_WINNERS: "최종 우승자",
+});
+
+export { INPUT_MESSAGE, OUTPUT_MESSAGE };
