+++ src/view/InputView.js
@@ -1,13 +1,8 @@
 import { Console } from "@woowacourse/mission-utils";
 
+import { INPUT_MESSAGE } from "../constants/io-message.js";
 import validateRacingCarNames from "../validator/car-validator.js";
 import validateAttemptCount from "../validator/attempt-validator.js";
-
-const INPUT_MESSAGE = Object.freeze({
-  RACING_CARS:
-    "경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)\n",
-  ATTEMPT_COUNT: "시도할 횟수는 몇 회인가요?\n",
-});
 
 const InputView = {
   /**
